+++ backend/src/main/java/woowacourse/touroot/travelogue/domain/Travelogue.java
@@ -5,11 +5,8 @@ import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.OneToMany;
-import java.util.List;
 import lombok.Getter;
 import woowacourse.touroot.entity.BaseEntity;
-import woowacourse.touroot.travelogue.day.domain.TravelogueDay;
 
 @Getter
 @Entity
@@ -24,7 +21,4 @@ public class Travelogue extends BaseEntity {
 
     @Column(nullable = false)
     private String thumbnail;
-
-    @OneToMany(mappedBy = "travelogue")
-    private List<TravelogueDay> travelogueDays;
 }
