+++ backend/src/main/java/woowacourse/touroot/travelogue/domain/day/dto/TravelogueDayResponse.java
@@ -1,4 +1,4 @@
-package woowacourse.touroot.travelogue.day.dto;
+package woowacourse.touroot.travelogue.domain.day.dto;
 
 import java.util.List;
 import woowacourse.touroot.travelogue.place.dto.TraveloguePlaceResponse;
