+++ backend/src/main/java/woowacourse/touroot/travelogue/domain/photo/domain/TraveloguePhoto.java
@@ -1,4 +1,4 @@
-package woowacourse.touroot.travelogue.day.domain;
+package woowacourse.touroot.travelogue.domain.photo.domain;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -8,28 +8,25 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
-import jakarta.persistence.OneToMany;
-import java.util.List;
 import lombok.Getter;
 import woowacourse.touroot.entity.BaseEntity;
-import woowacourse.touroot.travelogue.domain.Travelogue;
-import woowacourse.touroot.travelogue.place.domain.TraveloguePlace;
+import woowacourse.touroot.travelogue.domain.place.domain.TraveloguePlace;
 
 @Getter
 @Entity
-public class TravelogueDay extends BaseEntity {
+public class TraveloguePhoto extends BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @Column(name = "DAY_ORDER", nullable = false)
+    @Column(name = "PHOTO_KEY", nullable = false)
+    private String key;
+
+    @Column(name = "PHOTO_ORDER", nullable = false)
     private Integer order;
 
     @JoinColumn(nullable = false)
     @ManyToOne(fetch = FetchType.LAZY)
-    private Travelogue travelogue;
-
-    @OneToMany(mappedBy = "travelogueDay")
-    private List<TraveloguePlace> traveloguePlaces;
+    private TraveloguePlace traveloguePlace;
 }
