+++ backend/src/main/java/woowacourse/touroot/travelogue/domain/photo/repository/TraveloguePhotoRepository.java
@@ -1,7 +1,9 @@
-package woowacourse.touroot.travelogue.photo.repository;
+package woowacourse.touroot.travelogue.domain.photo.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import woowacourse.touroot.travelogue.photo.domain.TraveloguePhoto;
+import woowacourse.touroot.travelogue.domain.photo.domain.TraveloguePhoto;
+import woowacourse.touroot.travelogue.domain.place.domain.TraveloguePlace;
 
 public interface TraveloguePhotoRepository extends JpaRepository<TraveloguePhoto, Long> {
 }
