+++ backend/src/main/java/woowacourse/touroot/travelogue/domain/place/domain/TraveloguePlace.java
@@ -1,4 +1,4 @@
-package woowacourse.touroot.travelogue.place.domain;
+package woowacourse.touroot.travelogue.domain.place.domain;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -13,8 +13,7 @@ import java.util.List;
 import lombok.Getter;
 import woowacourse.touroot.entity.BaseEntity;
 import woowacourse.touroot.place.domain.Place;
-import woowacourse.touroot.travelogue.day.domain.TravelogueDay;
-import woowacourse.touroot.travelogue.photo.domain.TraveloguePhoto;
+import woowacourse.touroot.travelogue.domain.day.domain.TravelogueDay;
 
 @Getter
 @Entity
