--- backend/src/main/java/woowacourse/touroot/travelogue/place/repsitory/TraveloguePlaceRepository.java
@@ -1,7 +1,0 @@
-package woowacourse.touroot.travelogue.place.repsitory;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import woowacourse.touroot.travelogue.place.domain.TraveloguePlace;
-
-public interface TraveloguePlaceRepository extends JpaRepository<TraveloguePlace, Long> {
-}
